requirements.txt
streamlit
google-generativeai
tavily-python
PyPDF2
python-dotenv